Add sidechain audio effect

+++ ksmaudio/src/audio_effect/dsp/retrigger_echo_dsp.cpp
@@ -56,7 +56,7 @@ namespace ksmaudio::AudioEffect
         // トリガ更新までのフレーム数を計算
         if (params.secUntilTrigger >= 0.0f) // 負の値は無視
         {
-            const std::ptrdiff_t newFramesUntilTrigger = static_cast<std::ptrdiff_t>(params.secUntilTrigger * static_cast<float>(m_info.sampleRate));
+            const std::ptrdiff_t newFramesUntilTrigger = static_cast<std::ptrdiff_t>(params.secUntilTrigger * m_info.sampleRate);
 
             // 前回より小さい場合のみ反映(トリガ発生寸前に次の時間が入ることでトリガが抜ける現象を回避するため)
             if (m_framesUntilTrigger < 0 || m_framesUntilTrigger > newFramesUntilTrigger)
